FIB: fix bubble location, make white with dark bg
This was going off the right side of the page. I thought right was
supposed to position based off the right side of the div, but something
seems to be interfering with it working that way. As long as we
compensate for the width with the right positioning it works. Kind of
janky but it works.